wip: start extracting to separate crate
