+++ crates/script/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+name = "forge-script"
+description = "Solidity scripting"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+forge-verify.workspace = true
+foundry-cli.workspace = true
+foundry-config.workspace = true
+foundry-common.workspace = true
+foundry-evm.workspace = true
+foundry-debugger.workspace = true
+foundry-cheatcodes.workspace = true
+foundry-wallets.workspace = true
+
+hex.workspace = true
+serde.workspace = true
+eyre.workspace = true
+serde_json.workspace = true
+thiserror = "1"
+dunce = "1"
+foundry-compilers = { workspace = true, features = ["full"] }
+tracing.workspace = true
+clap = { version = "4", features = ["derive", "env", "unicode", "wrap_help"] }
+semver = "1"
+futures = "0.3"
+async-recursion = "1.0.5"
+alloy-primitives.workspace = true
+alloy-dyn-abi.workspace = true
+itertools.workspace = true
+parking_lot = "0.12"
+yansi = "0.5"
+ethers-core.workspace = true
+ethers-providers.workspace = true
+ethers-signers.workspace = true
+revm-inspectors.workspace = true
+alloy-rpc-types.workspace = true
+alloy-json-abi.workspace = true
+dialoguer = { version = "0.11", default-features = false }
+indicatif = "0.17"
