+++ crates/script/src/artifacts.rs
