+++ crates/script/src/build.rs
