+++ crates/script/src/resume.rs
