+++ crates/script/src/verify.rs
